+++ src/api/c/include/sqlrelay/sqlrclientwrapper.h
@@ -443,7 +443,7 @@ void	sqlrcur_inputBindDate(sqlrcur sqlrcurref,
 				const char *variable,
 				int16_t year, int16_t month, int16_t day,
 				int16_t hour, int16_t minute, int16_t second,
-				int16_t microsecond, const char *tz);
+				int32_t microsecond, const char *tz);
 
 /** @ingroup sqlrclientwrapper
  *  Defines a binary lob input bind variable. */
@@ -596,7 +596,7 @@ int	sqlrcur_getOutputBindDate(sqlrcur sqlrcurref,
 				const char *variable,
 				int16_t *year, int16_t *month, int16_t *day,
 				int16_t *hour, int16_t *minute, int16_t *second,
-				int16_t *microsecond, const char **tz);
+				int32_t *microsecond, const char **tz);
 
 /** @ingroup sqlrclientwrapper
  *  Get the value stored in a previously defined
