+++ src/api/cs/SQLRClient/SQLRelayCommand.cs
@@ -566,7 +566,7 @@ namespace SQLRClient
                             Int16 hour = 0;
                             Int16 minute = 0;
                             Int16 second = 0;
-                            Int16 microsecond = 0;
+                            Int32 microsecond = 0;
                             String tz = null;
                             _sqlrcur.getOutputBindDate(param.ParameterName, out year, out month, out day, out hour, out minute, out second, out microsecond, out tz);
                             param.Value = new DateTime(year, month, day, hour, minute, second, microsecond/1000);
