+++ src/connections/freetds/freetdsconnection.h
@@ -28,7 +28,7 @@ struct datebind {
         int16_t         *hour;
         int16_t         *minute;
         int16_t         *second;
-        int16_t         *microsecond;
+        int32_t         *microsecond;
         const char      **tz;
 };
 
@@ -66,7 +66,7 @@ class freetdscursor : public sqlrcursor_svr {
 						int16_t hour,
 						int16_t minute,
 						int16_t second,
-						int16_t microsecond,
+						int32_t microsecond,
 						const char *tz,
 						char *buffer,
 						uint16_t buffersize,
@@ -94,7 +94,7 @@ class freetdscursor : public sqlrcursor_svr {
 						int16_t *hour,
 						int16_t *minute,
 						int16_t *second,
-						int16_t *microsecond,
+						int32_t *microsecond,
 						const char **tz,
 						char *buffer,
 						uint16_t buffersize,
