+++ src/connections/odbc/odbcconnection.h
@@ -55,7 +55,7 @@ struct datebind {
 	int16_t		*hour;
 	int16_t		*minute;
 	int16_t		*second;
-	int16_t		*microsecond;
+	int32_t		*microsecond;
 	const char	**tz;
 	char		*buffer;
 };
@@ -90,7 +90,7 @@ class odbccursor : public sqlrcursor_svr {
 						int16_t hour,
 						int16_t minute,
 						int16_t second,
-						int16_t microsecond,
+						int32_t microsecond,
 						const char *tz,
 						char *buffer,
 						uint16_t buffersize,
@@ -118,7 +118,7 @@ class odbccursor : public sqlrcursor_svr {
 						int16_t *hour,
 						int16_t *minute,
 						int16_t *second,
-						int16_t *microsecond,
+						int32_t *microsecond,
 						const char **tz,
 						char *buffer,
 						uint16_t buffersize,
