+++ src/connections/router/routerconnection.h
@@ -24,7 +24,7 @@ struct outputbindvar {
 			int16_t		*hour;
 			int16_t		*minute;
 			int16_t		*second;
-			int16_t		*microsecond;
+			int32_t		*microsecond;
 			const char	**tz;
 		} datevalue;
 	} value;
@@ -70,7 +70,7 @@ class routercursor : public sqlrcursor_svr {
 						int16_t hour,
 						int16_t minute,
 						int16_t second,
-						int16_t microsecond,
+						int32_t microsecond,
 						const char *tz,
 						char *buffer,
 						uint16_t buffersize,
@@ -108,7 +108,7 @@ class routercursor : public sqlrcursor_svr {
 						int16_t *hour,
 						int16_t *minute,
 						int16_t *second,
-						int16_t *microsecond,
+						int32_t *microsecond,
 						const char **tz,
 						char *buffer,
 						uint16_t buffersize,
